If a wildcard line is followed by content, make sure it's matched.
Previously this test did not pass:

  assert!(!helper("a\n...\nb", "a"));

because as soon as we saw the "..." in the pattern *and* that we had reached the
end of the text, we assumed that the rest of the text must match, ignoring the
fact that there might still be more of the pattern to match. This commit fixes
that.