Show ERROR_CONTEXT lines of context either side of the error.
Previously we printed out the entire pattern and text when we encountered an
error. This is fine for tiny strings, but overwhelming for even moderately large
strings. This commit prints at most ERROR_CONTEXT lines either side of the error
line for example:

  Pattern (error at line 1000):
     ...
     |a997
     |a998
     |a999
  >> |a1000
     |a1001
     |a1002
     |a1003
     ...

  Text (error at line 1000):
     ...
     |a997
     |a998
     |a999
  >> |a1001
     |a1002
     |a1003
     |a1004
     ...

I chose ERROR_CONTEXT=3, because that means that the output is few enough lines
to be easily seen with a standard xterm's height. We could make this
configurable at some point if that seems useful, but it seems like overkill for
now.

Note that inputs that are small enough in length still display as they did
previously: this only effects longer inputs.