Allow multiple name matchers.
This allows you to specify multiple pattern regex's if you need them. It's
probably not going to be used by many people, but it's not hard to imagine it
being useful sometimes, and it nicely generalises the API.

This does imply an API change: you can no longer unset a name matcher. Honestly,
I can't imagine anyone ever wanted to do that. However, that does mean that this
feature will imply a bump to 0.2.