Implement grouping.
Previously, the following pattern:

```
...
a
b
```

matched as soon as "a" was found, so there was no way
to express "I want to match an 'a' only if it's immediately followed by
a 'b'". This commit maintains that behaviour but allows grouping. With
this commit one can write:

```
...
(((
a
b
)))
```

to express this. In other words `(((\n...\n)))` "groups" together the
lines inside and only matches if all the lines in there match. If it
doesn't, it advances the search by one line and tries again.

I suppose it was inevitable at some point, but this turns fm into a
limited, bounded, backtracking searcher. Whereas before the search was
best case and worst case _O(n)_ (where *n* was the number of lines in
the text (i.e. non-pattern)) the search is now best case _O(n)_ and
worst case _O(n * m)_ (where *m* is the number of lines in the pattern).