Fix matching with multiple name matchers.
This was very subtly wrong before, and could skip input and names,
causing incorrect strings to match (see the new tests for what didn't
work properly before).

The fix in this commit slows intra-line matching down considerably, as
we now check each character individually when name matchers are used. If
this turns out to be an actual bottleneck, this can be optimised.