Fix pattern error reporting bug.
When using the `..~` operator, it could eventually match after having
failed several times, and accidentally increment the apparent line the
pattern failed to match at. This didn't affect the matching algorithm,
but meant that a failure to match would report the wrong position in the
pattern. The new test line below incorrectly reported `(8, 6)` before
the fix in this commit.